move test and add arguments
