+++ tests/models/test_ProjectionHeads.py
@@ -93,23 +93,29 @@ class TestProjectionHeads(unittest.TestCase):
                         self.assertEqual(y.shape[0], batch_size)
                         self.assertEqual(y.shape[1], n_prototypes)
 
-    def test_swav_frozen_prototypes(self, device: str = 'cpu', seed=0):
-        for in_features, _, n_prototypes in self.n_features:
-            torch.manual_seed(seed)
-            prototypes = SwaVPrototypes(in_features, n_prototypes, 3)
-            prototypes = prototypes.eval()
-            prototypes = prototypes.to(device)
-            for batch_size in [1, 2]:
-                msg = 'prototypes d_in, n_prototypes = ' +\
-                    f'{in_features} x {n_prototypes}'
-                with self.subTest(msg=msg):
-                        x = torch.torch.rand((batch_size, in_features)).to(device)
-                        with torch.no_grad():
-                            y = prototypes(x, 1)
-                        self.assertEqual(y.shape[0], batch_size)
-                        self.assertEqual(y.shape[1], n_prototypes)
+    def test_swav_frozen_prototypes(self, device: str = "cpu", seed=0):
+        criterion = torch.nn.L1Loss()
+        linear_layer = torch.nn.Linear(8, 8, bias=False)
+        prototypes = SwaVPrototypes(input_dim=8, n_prototypes=8, n_steps_frozen_prototypes=2)
+        optimizer = torch.optim.SGD(prototypes.parameters(), lr=0.01)
+        torch.manual_seed(seed)
+        in_features = torch.rand(4, 8, device="cpu")
+        target_features = torch.ones(4, 8, device="cpu")
+        for step in range(4):
+            out_features = linear_layer(in_features)
+            out_features = prototypes.forward(out_features, step)
+            loss = criterion(out_features, target_features)
+            loss.backward()
+            optimizer.step()
+            optimizer.zero_grad()
+            if step == 0:
+                loss0 = loss
+            if step <= 2:
+                self.assertEqual(loss, loss0)
+            if step > 2:
+                self.assertNotEqual(loss, loss0) 
     
-    def test_swav_mutli_prototypes(self, device: str = "cpu", seed=0):
+    def test_swav_multi_prototypes(self, device: str = "cpu", seed=0):
         for in_features, _, n_prototypes in self.swavProtoypes:
             torch.manual_seed(seed)
             prototypes = SwaVPrototypes(in_features, n_prototypes)
